requirements.txt
Flask==2.1.2
opencv-python>=3.4.2
pillow>=7.1.0
torch>=0.4.1
torchvision>=0.2.0
numpy>=1.14.3
scipy>=1.0.0
h5py>=2.7.1
imageio>=2.4.1
pandas>=0.22.0
tqdm>=4.19.8
opencv-python>=3.4.2